% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Homework.R
\name{Gelman.Rubin}
\alias{Gelman.Rubin}
\title{Gelman-Rubin method of monitoring convergence}
\usage{
Gelman.Rubin(psi)
}
\arguments{
\item{psi}{A chain.}
}
\value{
The diagnostic statistics.
}
\description{
Use the Gelman-Rubin method to monitor convergence of the chain.
}
