% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MetropolisC}
\alias{MetropolisC}
\title{Metropolis sampler for Laplace distribution using Rcpp}
\usage{
MetropolisC(sigma, x0, N)
}
\arguments{
\item{sigma}{The Variance of normal distribution.}

\item{x0}{Initial value.}

\item{N}{The length of the chain.}
}
\value{
A chain.
}
\description{
Implement a random walk Metropolis sampler for generating the standard Laplace distribution.
}
