% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_functions.R
\name{relweights}
\alias{relweights}
\title{relative weight}
\usage{
relweights(fit)
}
\arguments{
\item{fit}{The fitted regression model.}
}
\value{
The relative weights of each predictor variable.
}
\description{
Compute the relative weights of each predictor variable of the regression model.
}
\examples{
\dontrun{
fit <- lm(z ~ x+y)
relweights(fit)
}
}
