% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Homework.R
\name{rw.Metropolis}
\alias{rw.Metropolis}
\title{Metropolis sampler for Laplace distribution.}
\usage{
rw.Metropolis(sigma, x0, N)
}
\arguments{
\item{sigma}{The Variance of normal distribution.}

\item{x0}{Initial value.}

\item{N}{The length of the chain.}
}
\value{
A chain.
}
\description{
Implement a random walk Metropolis sampler for generating the standard Laplace distribution.
}
\examples{
\dontrun{
N <- 2000
sigma <- 0.5
x0 <- 25
rw1 <- rw.Metropolis(sigma, x0, N)
}
}
