% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_functions.R
\name{shrinkage}
\alias{shrinkage}
\title{R-square's k-fold cross validation.}
\usage{
shrinkage(fit, k)
}
\arguments{
\item{fit}{The fitted regression model.}

\item{k}{k-fold.}
}
\value{
The original R-square and the k-fold cross-validated R-square.
}
\description{
The R-square 's k-fold cross validation is used to evaluate the generalization ability of regression model.
}
\examples{
\dontrun{
fit <- lm(z ~ x+y)
k <- 10
shrinkage(fit,k)
}
}
