% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Homework.R
\name{sk}
\alias{sk}
\title{Compute the sample skewness coeff using R.}
\usage{
sk(x)
}
\arguments{
\item{x}{The sample we get.}
}
\value{
the sample skewness coeff
}
\description{
Compute the sample skewness coeff using R.
}
\examples{
\dontrun{
x<-runif(1000)
sk(x)
}
}
